requirements.txt
scrapy>=2.11.0
pymongo>=4.5.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
minio>=7.2.0
requests>=2.31.0
lxml>=4.9.0
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.1.0
